Added both of the required Optimization algorithms.
Added a method that gets the value of the specified pixel in the
SLMPixelArray.

Added three methods in the SLMRegroupement.

The first is a private method that displays the same value on all of the
pixels in the first group on a given array.

The other two are optimization methods. The finds the best value for
every array in the regroupement by trying every value one by one (every
value is picked at a constant increment) The second arbitrarily selects
groups and changes them by a constant amount. If the change is helpful,
it is kept. If not, it is ignored.