Major Changes to Data Structures
Made some drastic changes to the data structures currently in the
program.

Completely removed the bit inside the Pixels that was used for selection
because I could simply not think of any good reason to keep it. Any
functionality it could provide could easily be replaced by other, more
memory efficient means.

Added TypeVerifiable class that now contains the methods for verifying
if a number is numeric and if a number is a whole number. All classes
are now children of TypeVerifiable and TypeVerifiable is a subclass of
handle.

Began removing useless methods in SLMPixelArray, SLMPixelGroup and
SLMRegroupement